Restructure solution; add Explanations module & console UI (#38)
- Split solution into MathsEngine.Models, MathsEngine.Console, and MathsEngine.Tests projects
- Introduced Explanations module with step-by-step "Tutor" classes for Pure, Mechanics, and Statistics
- Added CalculationResult class to encapsulate results and explanation steps
- Implemented new calculation modules for mechanics, pure math, and statistics
- Added comprehensive custom exception handling
- Built interactive console UI with menus for all major topics
- Added utility classes for input parsing and error display
- Updated tests to use double for bivariate analysis and reference new structure
- Added detailed README for Explanations module and updated solution file
- Improved error messages, input validation, and edge case handling throughout

+++ MathsEngine.Tests/ExplanationsTests/StatisticsTests/BivariateAnalysisTutorTests.cs
@@ -10,8 +10,8 @@ public class BivariateAnalysisTutorTests
     public void CalculateSpearmanRankWithSteps_PerfectPositiveCorrelation_ReturnsOne()
     {
         // Arrange
-        var scores1 = new List<int> { 10, 20, 30, 40, 50 };
-        var scores2 = new List<int> { 15, 25, 35, 45, 55 };
+        var scores1 = new List<double> { 10, 20, 30, 40, 50 };
+        var scores2 = new List<double> { 15, 25, 35, 45, 55 };
 
         // Act
         var result = BivariateAnalysisTutor.CalculateSpearmanRankWithSteps(scores1, scores2);
@@ -25,8 +25,8 @@ public class BivariateAnalysisTutorTests
     public void CalculateSpearmanRankWithSteps_PerfectNegativeCorrelation_ReturnsNegativeOne()
     {
         // Arrange
-        var scores1 = new List<int> { 10, 20, 30, 40, 50 };
-        var scores2 = new List<int> { 50, 40, 30, 20, 10 };
+        var scores1 = new List<double> { 10, 20, 30, 40, 50 };
+        var scores2 = new List<double> { 50, 40, 30, 20, 10 };
 
         // Act
         var result = BivariateAnalysisTutor.CalculateSpearmanRankWithSteps(scores1, scores2);
@@ -40,8 +40,8 @@ public class BivariateAnalysisTutorTests
     public void CalculateSpearmanRankWithSteps_GeneratesSteps()
     {
         // Arrange
-        var scores1 = new List<int> { 10, 20, 30 };
-        var scores2 = new List<int> { 15, 25, 35 };
+        var scores1 = new List<double> { 10, 20, 30 };
+        var scores2 = new List<double> { 15, 25, 35 };
 
         // Act
         var result = BivariateAnalysisTutor.CalculateSpearmanRankWithSteps(scores1, scores2);
@@ -59,8 +59,8 @@ public class BivariateAnalysisTutorTests
     public void CalculateSpearmanRankWithSteps_IncludesRankingDetails()
     {
         // Arrange
-        var scores1 = new List<int> { 5, 10, 15 };
-        var scores2 = new List<int> { 8, 12, 16 };
+        var scores1 = new List<double> { 5, 10, 15 };
+        var scores2 = new List<double> { 8, 12, 16 };
 
         // Act
         var result = BivariateAnalysisTutor.CalculateSpearmanRankWithSteps(scores1, scores2);
@@ -75,8 +75,8 @@ public class BivariateAnalysisTutorTests
     public void CalculateSpearmanRankWithSteps_ShowsDifferenceCalculations()
     {
         // Arrange
-        var scores1 = new List<int> { 1, 2, 3 };
-        var scores2 = new List<int> { 3, 2, 1 };
+        var scores1 = new List<double> { 1, 2, 3 };
+        var scores2 = new List<double> { 3, 2, 1 };
 
         // Act
         var result = BivariateAnalysisTutor.CalculateSpearmanRankWithSteps(scores1, scores2);
@@ -91,8 +91,8 @@ public class BivariateAnalysisTutorTests
     public void CalculateSpearmanRankWithSteps_IncludesFormula()
     {
         // Arrange
-        var scores1 = new List<int> { 10, 20, 30 };
-        var scores2 = new List<int> { 12, 22, 32 };
+        var scores1 = new List<double> { 10, 20, 30 };
+        var scores2 = new List<double> { 12, 22, 32 };
 
         // Act
         var result = BivariateAnalysisTutor.CalculateSpearmanRankWithSteps(scores1, scores2);
@@ -107,8 +107,8 @@ public class BivariateAnalysisTutorTests
     public void CalculateSpearmanRankWithSteps_IncludesInterpretation()
     {
         // Arrange
-        var scores1 = new List<int> { 10, 20, 30, 40 };
-        var scores2 = new List<int> { 15, 25, 35, 45 };
+        var scores1 = new List<double> { 10, 20, 30, 40 };
+        var scores2 = new List<double> { 15, 25, 35, 45 };
 
         // Act
         var result = BivariateAnalysisTutor.CalculateSpearmanRankWithSteps(scores1, scores2);
@@ -123,8 +123,8 @@ public class BivariateAnalysisTutorTests
     public void CalculateSpearmanRankWithSteps_WeakCorrelation()
     {
         // Arrange - specific values
-        var scores1 = new List<int> { 1, 2, 3, 4, 5 };
-        var scores2 = new List<int> { 3, 1, 5, 2, 4 };
+        var scores1 = new List<double> { 1, 2, 3, 4, 5 };
+        var scores2 = new List<double> { 3, 1, 5, 2, 4 };
 
         // Act
         var result = BivariateAnalysisTutor.CalculateSpearmanRankWithSteps(scores1, scores2);
